+++ R/ci.R
@@ -64,7 +64,7 @@ check_uses_travis <- function(base_path = proj_get()) {
 
   ui_stop(
     "
-    Cannot detect that package {ui_(project_name(base_path))} already uses Travis.
+    Cannot detect that package {ui_value(project_name(base_path))} already uses Travis.
     Do you need to run {ui_code('use_travis()')}?
     "
   )
@@ -142,7 +142,7 @@ check_uses_gitlab_ci <- function(base_path = proj_get()) {
 
   ui_stop(
     "
-    Cannot detect that package {ui_(project_name(base_path))} already uses GitLab CI.
+    Cannot detect that package {ui_value(project_name(base_path))} already uses GitLab CI.
     Do you need to run {ui_code('use_gitlab_ci()')}?
     "
   )