+++ R/course.R
@@ -1,45 +1,34 @@
 ## see end of file for some cURL notes
 
-#' Download course materials
+#' Download and unpack a ZIP file
 #'
-#' Special-purpose function to download a folder of course materials. The only
-#' demand on the user is to confirm or specify where the new folder should be
-#' stored. Workflow:
-#' * User executes something like: `use_course("bit.ly/xxx-yyy-zzz")`.
-#' * User is asked to notice and confirm the location of the new folder. Specify
-#' `destdir` to skip this.
-#' * User is asked if they'd like to delete the ZIP file.
-#' * If new folder contains an `.Rproj` file, it is opened. Otherwise, the
-#' folder is opened in the file manager, e.g. Finder or File Explorer.
+#' Functions to download and unpack a ZIP file into a local folder of files,
+#' with very intentional default behaviour. Useful in pedagogical settings or
+#' anytime you need a large audience to download a set of files quickly and
+#' actually be able to find them.
 #'
-#' If `url` has no "http" prefix, "https://" is prepended, allowing for even
-#' less typing by the user. A `url` of the form `"repo/user"`, is assumed to
-#' refer to a GitHub repository and is expanded appropriately, unless `url`
-#' matches a URL shortener service, currently only `"bit.ly"` or `"rstd.io"`.
-#' Most URL shorteners give HTTPS links and, anecdotally, we note this appears
-#' to work with [bit.ly](https://bitly.com/) links, even though they are
-#' nominally HTTP.
+#' @param url Link to a ZIP file containing the materials. Various short forms
+#'   are accepted, to reduce the typing burden in live settings:
 #'
-#' @param url Link to a ZIP file containing the materials, possibly behind a
-#'   shortlink. Links of the form `"repo/user"` are expanded to the appropriate
-#'   url for a repo on GitHub. Function developed with DropBox and GitHub in
-#'   mind, but should work for ZIP files generally. If no "http" prefix is
-#'   found, "https://" is prepended. See [use_course_details] for more.
-#' @param destdir The new folder is stored here. Defaults to user's Desktop.
+#'     * bit.ly or rstd.io shortlinks: "bit.ly/xxx-yyy-zzz" or "rstd.io/foofy"
+#'     * GitHub repo spec: "OWNER/REPO"
+#'   Function works well with DropBox folders and GitHub repos, but should work
+#'   for ZIP files generally. See examples and [use_course_details] for more.
+#' @param destdir The new folder is stored here. If `NULL`, defaults to user's
+#'   Desktop or some other conspicuous place.
+#' @param cleanup Whether to delete the original ZIP file after unpacking its
+#'   contents. In an interactive setting, `NA` leads to a menu where user can
+#'   approve the deletion (or decline).
 #'
-#' @return Path to the new directory holding the course materials, invisibly.
-#' @export
-#' @family download functions
+#' @return Path to the new directory holding the unpacked ZIP file, invisibly.
+#' @name zip-utils
 #' @examples
 #' \dontrun{
-#' ## bit.ly shortlink example
-#' ## should work with and without http prefix
+#' # download the source of usethis from GitHub, behind a bit.ly shortlink
 #' use_course("bit.ly/usethis-shortlink-example")
 #' use_course("http://bit.ly/usethis-shortlink-example")
 #'
-#' ## demo with a small CRAN package available in various places
-#'
-#' ## from CRAN
+#' ## download the source of rematch2 package, from CRAN and GitHub
 #' use_course("https://cran.r-project.org/bin/windows/contrib/3.4/rematch2_2.0.1.zip")
 #'
 #' ## from GitHub, 3 ways
@@ -47,38 +36,77 @@
 #' use_course("https://github.com/r-lib/rematch2/archive/master.zip")
 #' use_course("https://api.github.com/repos/r-lib/rematch2/zipball/master")
 #' }
+NULL
+
+#' @describeIn zip-utils Designed with live workshops in mind. Includes
+#'   intentional friction to highlight the download destination. Workflow:
+#' * User executes, e.g., `use_course("bit.ly/xxx-yyy-zzz")`.
+#' * User is asked to notice and confirm the location of the new folder. Specify
+#'   `destdir` to prevent this.
+#' * User is asked if they'd like to delete the ZIP file.
+#' * If new folder contains an `.Rproj` file, a new instance of RStudio is
+#'   launched. Otherwise, the folder is opened in the file manager, e.g. Finder
+#'   or File Explorer.
+#' @export
 use_course <- function(url, destdir = NULL) {
   url <- normalize_url(url)
-  zipfile <- download_zip(
-    url,
-    destdir = destdir %||% conspicuous_place(),
-    pedantic = is.null(destdir) && interactive()
-  )
-  tidy_unzip(zipfile)
+  destdir_not_specified <- is.null(destdir)
+  destdir <- user_path_prep(destdir %||% conspicuous_place())
+  check_path_is_directory(destdir)
+
+  if (destdir_not_specified && interactive()) {
+    ui_line(c(
+      "Downloading into {ui_path(destdir)}.",
+      "Prefer a different location? Cancel, try again, and specify {ui_code('destdir')}"
+    ))
+    if (ui_nope("OK to proceed?")) {
+      ui_stop("Aborting.")
+    }
+  }
+
+  ui_done("Downloading from {ui_value(url)}")
+  zipfile <- tidy_download(url, destdir)
+  ui_done("Download stored in {ui_path(zipfile)}")
+  check_is_zip(attr(zipfile, "content-type"))
+  tidy_unzip(zipfile, cleanup = NA)
 }
 
-#' Download and unpack a ZIP file
+#' @describeIn zip-utils More useful in day-to-day work. Downloads in current
+#'   working directory, by default, and allows `cleanup` behaviour to be
+#'   specified.
+#' @export
+use_zip <- function(url,
+                    destdir = getwd(),
+                    cleanup = if (interactive()) NA else FALSE) {
+  url <- normalize_url(url)
+  check_path_is_directory(destdir)
+  ui_done("Downloading from {ui_value(url)}")
+  zipfile <- tidy_download(url, destdir)
+  ui_done("Download stored in {ui_path(zipfile)}")
+  check_is_zip(attr(zipfile, "content-type"))
+  tidy_unzip(zipfile, cleanup)
+}
+
+#' Helpers to download and unpack a ZIP file
 #'
-#' Details on the two functions that power [use_course()]. These internal
-#' functions are currently unexported but a course instructor may want more
+#' Details on the internal functions that power [use_course()] and [use_zip()].
+#' These helpers are currently unexported but a course instructor may want more
 #' details.
 #'
 #' @name use_course_details
-#' @family download functions
 #' @keywords internal
 #'
-#' @section download_zip():
+#' @section tidy_download():
 #'
 #' ```
 #' ## function signature
-#' download_zip(url, destdir = getwd(), pedantic = FALSE)
+#' tidy_download(url, destdir = getwd())
 #'
 #' ## as called inside use_course()
-#' download_zip(
+#' tidy_download(
 #'   url, ## after post-processing with normalize_url()
-#'   ## conspicuous_place() = Desktop or home directory or working directory
-#'   destdir = destdir %||% conspicuous_place(),
-#'   pedantic = is.null(destdir) && interactive()
+#'   # conspicuous_place() = Desktop or home directory or working directory
+#'   destdir = destdir %||% conspicuous_place()
 #' )
 #' ```
 #'
@@ -106,7 +134,7 @@ use_course <- function(url, destdir = NULL) {
 #' https://www.dropbox.com/sh/12345abcde/6789wxyz?dl=1
 #' ```
 #' This download link (or a shortlink that points to it) is suitable as input
-#' for `download_zip()`. After one or more redirections, this link will
+#' for `tidy_download()`. After one or more redirections, this link will
 #' eventually lead to a download URL. For more details, see
 #' <https://www.dropbox.com/help/desktop-web/force-download> and
 #' <https://www.dropbox.com/en/help/desktop-web/download-entire-folders>.
@@ -119,7 +147,7 @@ use_course <- function(url, destdir = NULL) {
 #' https://github.com/r-lib/usethis/archive/master.zip
 #' ```
 #' This download link (or a shortlink that points to it) is suitable as input
-#' for `download_zip()`. After one or more redirections, this link will
+#' for `tidy_download()`. After one or more redirections, this link will
 #' eventually lead to a download URL. Here's an alternative link that also leads
 #' to ZIP download, albeit with a different filenaming scheme:
 #' ```
@@ -131,14 +159,10 @@ use_course <- function(url, destdir = NULL) {
 #' @param destdir Path to existing local directory where the ZIP file will be
 #'   stored. Defaults to current working directory, but note that [use_course()]
 #'   has different default behavior.
-#' @param pedantic Logical. When `TRUE` and in an interactive session, the user
-#'   is told where the ZIP file will be stored. If happy, user can elect to
-#'   proceed. Otherwise, user can abort and try again with the desired
-#'   `destdir`. Intentional friction for a pedagogical setting.
 #'
 #' @examples
 #' \dontrun{
-#' download_zip("https://github.com/r-lib/rematch2/archive/master.zip")
+#' tidy_download("https://github.com/r-lib/rematch2/archive/master.zip")
 #' }
 #'
 #' @section tidy_unzip():
@@ -170,50 +194,36 @@ use_course <- function(url, destdir = NULL) {
 #'
 #' @examples
 #' \dontrun{
-#' download_zip("https://github.com/r-lib/rematch2/archive/master.zip")
+#' tidy_download("https://github.com/r-lib/rematch2/archive/master.zip")
 #' tidy_unzip("rematch2-master.zip")
 #' }
 NULL
 
-download_zip <- function(url, destdir = getwd(), pedantic = FALSE) {
-  stopifnot(is_string(url))
-  base_path <- destdir
-  check_path_is_directory(base_path)
-
+# 1. downloads from `url`
+# 2. determines filename from content-description header (with fallbacks)
+# 3. returned path has content-type and content-description as attributes
+tidy_download <- function(url, destdir = getwd()) {
+  check_path_is_directory(destdir)
+  tmp <- file_temp("tidy-download-")
   h <- curl::new_handle(noprogress = FALSE, progressfunction = progress_fun)
-  tmp <- file_temp("usethis-use-course-")
-  ui_done("Downloading {ui_value(url)}")
-  curl::curl_download(
-    url, destfile = tmp, quiet = FALSE, mode = "wb", handle = h
-  )
-  check_is_zip(h)
+  curl::curl_download(url, tmp, quiet = FALSE, mode = "wb", handle = h)
   cat_line()
 
   cd <- content_disposition(h)
   base_name <- make_filename(cd, fallback = path_file(url))
-
-  ## DO YOU KNOW WHERE YOUR STUFF IS GOING?!?
-  if (interactive() && pedantic) {
-    ui_line(c(
-      "The ZIP file, {ui_value(base_name)}, will be copied to  {ui_path(base_path)}.",
-      "Prefer a different location? Cancel, try again, and specify {ui_code('destdir')}"
-    ))
-    if (ui_nope("OK to proceed?")) {
-      ui_stop("Aborting.")
-    }
-  }
-  full_path <- path(base_path, base_name)
+  full_path <- path(destdir, base_name)
 
   if (!can_overwrite(full_path)) {
     ui_stop("Aborting.")
   }
+  attr(full_path, "content-type") <- content_type(h)
+  attr(full_path, "content-disposition") <- cd
 
-  zip_dest <- if (is.null(destdir)) base_name else full_path
-  ui_done("Copied ZIP file to {ui_path(zip_dest, base_path)}")
   file_move(tmp, full_path)
+  invisible(full_path)
 }
 
-tidy_unzip <- function(zipfile) {
+tidy_unzip <- function(zipfile, cleanup = FALSE) {
   base_path <- path_dir(zipfile)
 
   filenames <- utils::unzip(zipfile, list = TRUE)[["Name"]]
@@ -239,12 +249,17 @@ tidy_unzip <- function(zipfile) {
     ({length(filenames)} files extracted)"
   )
 
-  if (interactive()) {
-    if (ui_yeah("Shall we delete the ZIP file ({ui_path(zipfile, base_path)})?")) {
-      ui_done("Deleting {ui_path(zipfile, base_path)}")
-      file_delete(zipfile)
-    }
+  if (isNA(cleanup)) {
+    cleanup <- interactive() &&
+      ui_yeah("Shall we delete the ZIP file ({ui_path(zipfile, base_path)})?")
+  }
 
+  if (isTRUE(cleanup)) {
+    ui_done("Deleting {ui_path(zipfile, base_path)}")
+    file_delete(zipfile)
+  }
+
+  if (interactive()) {
     rproj_path <- dir_ls(target, regexp = "[.]Rproj$")
     if (length(rproj_path) == 1 && rstudioapi::hasFun("openProject")) {
       ui_done("Opening project in RStudio")
@@ -316,17 +331,10 @@ top_directory <- function(filenames) {
   }
 }
 
-check_is_zip <- function(h) {
+content_type <- function(h) {
   headers <- curl::parse_headers_list(curl::handle_data(h)$headers)
-  if (headers[["content-type"]] != "application/zip") {
-    ui_stop(c(
-      "Download does not have MIME type {ui_value('application/zip')}.",
-      "Instead it's {ui_value(headers[['content-type']])}."
-    ))
-  }
-  invisible()
+  headers[["content-type"]]
 }
-
 
 content_disposition <- function(h) {
   headers <- curl::parse_headers_list(curl::handle_data(h)$headers)
@@ -337,10 +345,26 @@ content_disposition <- function(h) {
   parse_content_disposition(cd)
 }
 
+check_is_zip <- function(ct) {
+  # "https://www.fueleconomy.gov/feg/epadata/16data.zip" comes with
+  # MIME type "application/x-zip-compressed"
+  # see https://github.com/r-lib/usethis/issues/573
+  allowed <- c("application/zip", "application/x-zip-compressed")
+  if (!ct %in% allowed ) {
+    ui_stop(c(
+      "Download does not have MIME type {ui_value('application/zip')}.",
+      "Instead it's {ui_value(ct)}."
+    ))
+  }
+  invisible(ct)
+}
+
 ## https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Disposition
 ## https://tools.ietf.org/html/rfc6266
 ## DropBox eg: "attachment; filename=\"foo.zip\"; filename*=UTF-8''foo.zip\"
 ##  GitHub eg: "attachment; filename=foo-master.zip"
+# https://stackoverflow.com/questions/30193569/get-content-disposition-parameters
+# http://test.greenbytes.de/tech/tc2231/
 parse_content_disposition <- function(cd) {
   if (!grepl("^attachment;", cd)) {
     ui_stop(c(