+++ R/release.R
@@ -38,6 +38,7 @@ release_checklist <- function(version) {
   type <- release_type(version)
   on_cran <- !is.null(cran_version())
   has_src <- dir_exists(proj_path("src"))
+  has_news <- file_exists(proj_path("NEWS.md"))
 
   todo <- function(x, cond = TRUE) {
     x <- glue(x, .envir = parent.frame())
@@ -73,7 +74,7 @@ release_checklist <- function(version) {
     todo("Accepted :tada:"),
     todo("`usethis::use_github_release()`"),
     todo("`usethis::use_dev_version()`"),
-    todo("`usethis::use_news()`", !on_cran),
+    todo("`usethis::use_news_md()`", !has_news),
     todo("Update install instructions in README", !on_cran),
     todo("Finish blog post", type != "patch"),
     todo("Tweet"),