+++ tests/testthat/test-use-make.R
@@ -1,0 +1,7 @@
+context("use_makefile")
+
+test_that("use_make() creates a Makefile at project root", {
+  pkg <- scoped_temporary_package()
+  use_make()
+  expect_proj_file("Makefile")
+})